Refactored ketchup-buutton to kup-btn
