+++ packages/ketchup/src/components/kup-button/readme.md
@@ -1,4 +1,4 @@
-# ketchup-button
+# kup-button
