+++ packages/ketchup/src/index.html
@@ -64,7 +64,7 @@
         <br />
         <a href="data-table.html">Ketchup data table</a>
         <br />
-        <ketchup-button label="Sono bottone"></ketchup-button>
+        <kup-button label="Sono bottone"></kup-button>
         <ketchup-radio
             id="krg1"
             displayed-field="label"
